+++ CHANGELOG.md
@@ -1,8 +1,13 @@
 # (MODX)EvolutionCMS.snippets.ddGetDocumentField changelog
 
 
+## Version 2.10.4 (2020-09-29)
+* \* `\ddGetDocumentField\DataProvider\Document\DataProvider::get`: Fixed wrong outputting of alternative resource fields.
+* \+ README → Links.
+
+
 ## Version 2.10.3 (2020-07-05)
-* `\ddGetDocumentField\Input::paramsBackwardCompatibility`: Fixed wrong CMS event logging.
+* \* `\ddGetDocumentField\Input::paramsBackwardCompatibility`: Fixed wrong CMS event logging.
 
 
 ## Version 2.10.2 (2020-05-11)